CommentBox Role 100% Career Role 100% (No se si ponerle PUT o PATCH o ambas)

+++ src/main/java/org/example/conectatec/career/domain/CareerService.java
@@ -1,7 +1,9 @@
 package org.example.conectatec.career.domain;
 
 import jakarta.transaction.Transactional;
+import org.example.conectatec.auth.utils.AuthorizationUtils;
 import org.example.conectatec.career.infrastructure.CareerRepository;
+import org.example.conectatec.user.exceptions.UnauthorizeOperationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -10,9 +12,11 @@ import java.util.List;
 public class CareerService {
 
     private final CareerRepository careerRepository;
+    private final AuthorizationUtils authorizationUtils;
     @Autowired
-    public CareerService(CareerRepository careerRepository) {
+    public CareerService(CareerRepository careerRepository, AuthorizationUtils authorizationUtils) {
         this.careerRepository = careerRepository;
+        this.authorizationUtils = authorizationUtils;
     }
 
     @Transactional
@@ -27,11 +31,18 @@ public class CareerService {
 
     @Transactional
     public Career createCareer(Career career) {
+        String username = authorizationUtils.getCurrentUserEmail();
+        if(username == null){
+            throw new UnauthorizeOperationException("You do not have permission to access this resource");
+        }
         return careerRepository.save(career);
     }
 
     @Transactional
     public void deleteCareerById(Long id) {
+        if(!authorizationUtils.isAdminOrResourceOwner(id)){
+            throw new UnauthorizeOperationException("You do not have permission to access this resource");
+        }
         Career career = careerRepository.findById(id).get();
         careerRepository.delete(career);
     }
